refactor: [M3-6307] - MUIv5 Migration - SRC > Components > CopyTooltip
