+++ packages/manager/src/components/CopyTooltip/CopyTooltip.stories.mdx
@@ -1,22 +1,20 @@
 import { ArgsTable, Canvas, Meta, Story } from '@storybook/addon-docs';
-import CopyTooltip from 'src/components/CopyTooltip/CopyTooltip';
+import { CopyTooltip } from 'src/components/CopyTooltip/CopyTooltip';
 
 <Meta
   title="Elements/Tooltip/Copy Tooltip"
   component={CopyTooltip}
   decorators={[
     (Story) => (
-      <div style={{marginTop: '1.5em'}}>
+      <div style={{ marginTop: '1.5em' }}>
         <Story />
       </div>
-    )
+    ),
   ]}
 />
 
 export const Template = (args) => {
-  return (
-    <CopyTooltip {...args} />
-  );
+  return <CopyTooltip {...args} />;
 };
 
 # Copy Tooltip
@@ -42,4 +40,4 @@ If horizontal space is limited, truncate the content.
   </Story>
 </Canvas>
 
-<ArgsTable story="Copy Tooltip" exclude={['className']} sort="requiredFirst"/>
+<ArgsTable story="Copy Tooltip" exclude={['className']} sort="requiredFirst" />
