--- packages/manager/src/components/CopyTooltip/index.ts
@@ -1,2 +1,0 @@
-import CopyTooltip from './CopyTooltip';
-export default CopyTooltip;
