+++ packages/manager/src/components/CopyableAndDownloadableTextField.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import snakeCase from 'lodash/snakeCase';
-import CopyTooltip from 'src/components/CopyTooltip';
+import { CopyTooltip } from 'src/components/CopyTooltip/CopyTooltip';
 import DownloadTooltip from 'src/components/DownloadTooltip';
 import { makeStyles } from '@mui/styles';
 import { Theme } from '@mui/material/styles';
