+++ packages/manager/src/features/Longview/shared/InstallationInstructions.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import { makeStyles } from '@mui/styles';
 import { Theme } from '@mui/material/styles';
 
-import CopyTooltip from 'src/components/CopyTooltip';
+import { CopyTooltip } from 'src/components/CopyTooltip/CopyTooltip';
 import Typography from 'src/components/core/Typography';
 import Grid from 'src/components/Grid';
