+++ packages/manager/src/features/linodes/LinodesCreate/CodeBlock/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import HighlightedMarkdown from 'src/components/HighlightedMarkdown';
-import CopyTooltip from 'src/components/CopyTooltip';
+import { CopyTooltip } from 'src/components/CopyTooltip/CopyTooltip';
 import { sendEvent } from 'src/utilities/ga';
 import { useCodeBlockStyles } from './styles';
 export interface Props {
